Rename CreateSlide.vue to AuthorSlide.vue to better reflect the component's role.
