+++ src/components/AuthorSlide.vue
